IDCheck
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@3805 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/local_utl/showipdeny.c
@@ -6,6 +6,7 @@ int main(int argc, char **argv)
 {
     FILE *fp, *fp2;
     char fn[80],fn2[80];
+    char id[IDLEN+2];
     char * p;
     int i,j,ip[4],t,k,count=0;
     time_t tt;
@@ -34,20 +35,60 @@ int main(int argc, char **argv)
         unlink(fn);
         if(count==0) {
             unlink(fn2);
-            return;
         }
-        bzero(&deliveruser, sizeof(struct userec));
-        strcpy(deliveruser.userid, "deliver");
-        deliveruser.userlevel = -1;
-        strcpy(deliveruser.username, "自动发信系统");
-        currentuser = &deliveruser;
-        strcpy(fromhost, "天堂");
-        tt = time(0);
-        p = ctime(&tt);
-        p[19]=0; p+=4;
-        sprintf(fn, "%s的犯罪记录", p);
-        post_file(&deliveruser, NULL, fn2, "surrender", fn, 0, 1);
-        unlink(fn2);
+        else {
+            bzero(&deliveruser, sizeof(struct userec));
+            strcpy(deliveruser.userid, "deliver");
+            deliveruser.userlevel = -1;
+            strcpy(deliveruser.username, "自动发信系统");
+            currentuser = &deliveruser;
+            strcpy(fromhost, "天堂");
+            tt = time(0);
+            p = ctime(&tt);
+            p[19]=0; p+=4;
+            sprintf(fn, "%s 的 IP 犯罪记录", p);
+            post_file(&deliveruser, NULL, fn2, "surrender", fn, 0, 1);
+            unlink(fn2);
+        }
+    }
+
+    strcpy(fn, ".IDdenys");
+    strcpy(fn2, "tmp/showiddeny.txt");
+    fp=fopen(fn, "r");
+    if(fp) {
+        fp2=fopen(fn2, "w");
+        while(!feof(fp)) {
+            if(fscanf(fp, "%d %ld %s %d", &i, &j, id, &t)<=0) break;
+            if(t<=5) continue;
+            count++;
+            tt=(time_t) j;
+            p = ctime(&tt);
+            p[19]=0; p+=4;
+            if(i==0)
+                fprintf(fp2, "%s %s 两次连接时间太短.一小时内共连接%d次.\n", p, id,t+1);
+            else
+                fprintf(fp2, "%s %s 连接过于频繁.    一小时内共连接%d次.\n", p, id,t);
+        }
+        fclose(fp);
+        fclose(fp2);
+        unlink(fn);
+        if(count==0) {
+            unlink(fn2);
+        }
+        else {
+            bzero(&deliveruser, sizeof(struct userec));
+            strcpy(deliveruser.userid, "deliver");
+            deliveruser.userlevel = -1;
+            strcpy(deliveruser.username, "自动发信系统");
+            currentuser = &deliveruser;
+            strcpy(fromhost, "天堂");
+            tt = time(0);
+            p = ctime(&tt);
+            p[19]=0; p+=4;
+            sprintf(fn, "%s 的 ID 犯罪记录", p);
+            post_file(&deliveruser, NULL, fn2, "surrender", fn, 0, 1);
+            unlink(fn2);
+        }
     }
     return 0;
 }
